refactor into more reasonable namespaces
